feat: bookId 추가

+++ src/main/java/book/talk/createbook/model/BookRequest.java
@@ -6,5 +6,5 @@ import lombok.Data;
 public class BookRequest {
     private String thumbnail;
     private String bookTitle;
-    private int isbn13;
+    private Long isbn13;
 }
